+++ index.js
@@ -3,15 +3,12 @@ var path = require('path')
 var mkdirp = require('mkdirp')
 var Promise = require('rsvp').Promise
 var quickTemp = require('quick-temp')
-var Writer = require('broccoli-writer')
 var helpers = require('broccoli-kitchen-sink-helpers')
 var walkSync = require('walk-sync')
 var mapSeries = require('promise-map-series')
 
 
 module.exports = Filter
-Filter.prototype = Object.create(Writer.prototype)
-Filter.prototype.constructor = Filter
 function Filter (inputTree, options) {
   if (!inputTree) {
     throw new Error('broccoli-filter must be passed an inputTree, instead it received `undefined`');
@@ -24,34 +21,48 @@ function Filter (inputTree, options) {
   if (options.outputEncoding !== undefined) this.outputEncoding = options.outputEncoding
 }
 
-Filter.prototype.getCacheDir = function () {
-  return quickTemp.makeOrReuse(this, 'tmpCacheDir')
-}
-
-Filter.prototype.write = function (readTree, destDir) {
+Filter.prototype.rebuild = function () {
   var self = this
 
-  return readTree(this.inputTree).then(function (srcDir) {
-    var paths = walkSync(srcDir)
-
+  var paths = walkSync(this.inputPath)
     return mapSeries(paths, function (relativePath) {
       if (relativePath.slice(-1) === '/') {
-        mkdirp.sync(destDir + '/' + relativePath)
+        mkdirp.sync(self.outputPath + '/' + relativePath)
       } else {
         if (self.canProcessFile(relativePath)) {
-          return self.processAndCacheFile(srcDir, destDir, relativePath)
+          return self.processAndCacheFile(self.inputPath, self.outputPath, relativePath)
         } else {
           helpers.copyPreserveSync(
-            srcDir + '/' + relativePath, destDir + '/' + relativePath)
+            self.inputPath + '/' + relativePath, self.outputPath + '/' + relativePath)
         }
       }
     })
-  })
+}
+
+// Compatibility with Broccoli < 0.14
+// See https://github.com/broccolijs/broccoli/blob/master/docs/new-rebuild-api.md
+Filter.prototype.read = function (readTree) {
+  var self = this
+
+  quickTemp.makeOrRemake(this, 'outputPath')
+  quickTemp.makeOrReuse(this, 'cachePath')
+  this.needsCleanup = true
+
+  return readTree(this.inputTree)
+    .then(function (inputPath) {
+      self.inputPath = inputPath
+      return self.rebuild()
+    })
+    .then(function () {
+      return self.outputPath
+    })
 }
 
 Filter.prototype.cleanup = function () {
-  quickTemp.remove(this, 'tmpCacheDir')
-  Writer.prototype.cleanup.call(this)
+  if (this.needsCleanup) {
+    quickTemp.remove(this, 'outputPath')
+    quickTemp.remove(this, 'cachePath')
+  }
 }
 
 Filter.prototype.canProcessFile = function (relativePath) {
@@ -70,6 +81,9 @@ Filter.prototype.getDestFilePath = function (relativePath) {
   }
   return null
 }
+
+// To do: Get rid of the srcDir/destDir args because we now have inputPath/outputPath
+// https://github.com/search?q=processAndCacheFile&type=Code&utf8=%E2%9C%93
 
 Filter.prototype.processAndCacheFile = function (srcDir, destDir, relativePath) {
   var self = this
@@ -109,7 +123,7 @@ Filter.prototype.processAndCacheFile = function (srcDir, destDir, relativePath) 
       // the cache directory; we need to be 100% sure though that we don't try
       // to hardlink symlinks, as that can lead to directory hardlinks on OS X
       helpers.copyPreserveSync(
-        self.getCacheDir() + '/' + cacheEntry.cacheFiles[i], dest)
+        self.cachePath + '/' + cacheEntry.cacheFiles[i], dest)
     }
   }
 
@@ -124,7 +138,7 @@ Filter.prototype.processAndCacheFile = function (srcDir, destDir, relativePath) 
       cacheEntry.cacheFiles.push(cacheFile)
       helpers.copyPreserveSync(
         destDir + '/' + cacheEntry.outputFiles[i],
-        self.getCacheDir() + '/' + cacheFile)
+        self.cachePath + '/' + cacheFile)
     }
     cacheEntry.hash = hash(cacheEntry.inputFiles)
     self._cache[relativePath] = cacheEntry